% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_holidays.R, R/holiday_data.R
\name{add_holidays}
\alias{add_holidays}
\alias{USValentinesDay}
\alias{USSuperBowl}
\alias{USSaintPatricksDay}
\alias{USMothersDay}
\alias{USFathersDay}
\alias{USHalloweenDay}
\alias{USCyberMonday}
\alias{USBack2School}
\alias{USTgt_Cirle_Wk_Summer}
\alias{USTgt_Cirle_Wk_Fall}
\alias{USJuneteenthDay}
\alias{Holiday_Names}
\title{Added Holidays}
\usage{
USValentinesDay(year = timeDate::getRmetricsOptions("currentYear"))

USSuperBowl(year = timeDate::getRmetricsOptions("currentYear"))

USSaintPatricksDay(year = timeDate::getRmetricsOptions("currentYear"))

USMothersDay(year = timeDate::getRmetricsOptions("currentYear"))

USFathersDay(year = timeDate::getRmetricsOptions("currentYear"))

USHalloweenDay(year = timeDate::getRmetricsOptions("currentYear"))

USCyberMonday(year)

USBack2School(year = timeDate::getRmetricsOptions("currentYear"))

USTgt_Cirle_Wk_Summer(year = timeDate::getRmetricsOptions("currentYear"))

USTgt_Cirle_Wk_Fall(year = timeDate::getRmetricsOptions("currentYear"))

USJuneteenthDay(year = timeDate::getRmetricsOptions("currentYear"))

Holiday_Names(name, holiday_dates)
}
\arguments{
\item{year}{\code{numeric} A vector of year integers.}

\item{name}{\code{character} Name of the holiday represented by the holiday_dates vector.}

\item{holiday_dates}{\code{Dates} A vector of dates representing the holiday.}
}
\value{
These functions return the dates of the holiday specified in the
name of the function. For example, \code{USValentinesDay()} will provide the dates
for valentines day in the US when passed a vector of years.
}
\description{
These functions add holiday dates that are not found in the
timeDate package.  These are either new federal holidays, or they are retail
holidays which are not covered in the timeDate package.
}
\examples{
{
# Individual Holidays
library(holiday.dates)
years <- seq(2015,2020, by = 1)
US_ValentinesDay <- USValentinesDay(years) \%>\% as.Date()
Name_Valentines <- Holiday_Names("ValentinesDay",US_ValentinesDay)
}
}
