% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holiday_data.R
\name{holiday_data}
\alias{holiday_data}
\title{Holiday Data}
\usage{
holiday_data(df, out_df = c("weeks", "days"))
}
\arguments{
\item{df}{\code{dataframe} A dataframe with 1 date column with at least 2 dates.}

\item{out_df}{\code{character} \code{weeks} is the default, but can also use \code{days}.}
}
\value{
The result of the function can be either a dataframe with days between
the minimum and maximum of the dates provided, or, the default option which is
to return a dataframe with weeks instead of days.

The paramater \code{out_df} gives the user 2 choices for output.  The
default output a weeks dataframe with beginning week, ending week, a holiday
flag, and a column for the holiday name.  The second option returns a
dataframe with all the days and day names with the 4 columns mention above.
}
\description{
Builds a dataframe of holiday dates between 2 dates in the
form of 'yyyy-mm-dd'.  Pass a dataframe with the column of dates.
}
\note{
\itemize{
\item https://github.com/cran/timeDate/blob/master/R/holiday-Dates.R
\item https://www.federalpay.org/holidays
}
}
\examples{
\dontrun{
# Holiday dates in a dataframe
library(holiday.dates)
weeks_df <- tibble::tibble(weeks = seq.Date(as.Date('2015-01-01'),
as.Date('2020-01-01'), by = 'week'))
holidays_weeks(df = weeks_df) \%>\% head(10)
}

}
